requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
autopep8==2.3.2
beautifulsoup4==4.12.3
billiard==4.2.1
black==25.1.0
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
cron-descriptor==1.4.5
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.7
django-celery-beat==2.7.0
django-filter==25.1
django-silk==5.3.2
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gprof2dot==2024.6.6
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jiter==0.9.0
kombu==5.5.2
Markdown==3.7
mypy-extensions==1.0.0
mysqlclient==2.2.7
numpy==2.2.4
openai==1.74.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pytest==8.3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tiktoken==0.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
